requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.1.0
django-filter==23.2
django-jazzmin==2.6.0
django-simple-history==3.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docopt==0.6.2
et-xmlfile==1.1.0
Faker==19.2.0
idna==3.4
mysqlclient==2.1.1
num2words==0.5.12
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
